Add error details

+++ backend/src/main/java/com/employer/management/controller/EmployeeController.java
@@ -1,6 +1,7 @@
 package com.employer.management.controller;
 
 import com.employer.management.bean.Employee;
+import com.employer.management.exceptions.ResourceNotFoundException;
 import com.employer.management.repository.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -30,7 +31,7 @@ public class EmployeeController {
     }
 
     @GetMapping("/employees/{id}")
-    public ResponseEntity<Employee> getEmployeeById(@PathVariable(value = "id") Long employeeId)
+    public ResponseEntity<Employee> getEmployeeById(@PathVariable(value = "id") Integer employeeId)
             throws ResourceNotFoundException {
         Employee employee = employeeRepository.findById(employeeId)
                 .orElseThrow(() -> new ResourceNotFoundException("Employee not found for this id :: " + employeeId));
@@ -43,7 +44,7 @@ public class EmployeeController {
     }
 
     @PutMapping("/employees/{id}")
-    public ResponseEntity<Employee> updateEmployee(@PathVariable(value = "id") Long employeeId,
+    public ResponseEntity<Employee> updateEmployee(@PathVariable(value = "id") Integer employeeId,
             @Valid @RequestBody Employee employeeDetails) throws ResourceNotFoundException {
         Employee employee = employeeRepository.findById(employeeId)
                 .orElseThrow(() -> new ResourceNotFoundException("Employee not found for this id :: " + employeeId));
@@ -56,7 +57,7 @@ public class EmployeeController {
     }
 
     @DeleteMapping("/employees/{id}")
-    public Map<String, Boolean> deleteEmployee(@PathVariable(value = "id") Long employeeId)
+    public Map<String, Boolean> deleteEmployee(@PathVariable(value = "id") Integer employeeId)
             throws ResourceNotFoundException {
         Employee employee = employeeRepository.findById(employeeId)
                 .orElseThrow(() -> new ResourceNotFoundException("Employee not found for this id :: " + employeeId));
